requirements.txt
python-dotenv
requests
beautifulsoup4
anthropic
openpyxl
tweepy>=4.14.0

# Google integrations (uncomment as needed)
# google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
# gspread
